bonnes structures ouiiiiiiigit add .git add .

+++ etu/include/q_zz.h
@@ -5,7 +5,7 @@
 #include "qtables.h"
 #include "decoupage_mcu.h"
 
-extern void quantification_composante(int8_t *composante);
-extern int8_t* zigzag_composante(int8_t *composante);
+extern void quantification_composante(int16_t *composante);
+extern int16_t* zigzag_composante(int16_t *composante);
 
 #endif /* q_zz */
